+++ common/src/main/java/com/lne_rogues/LNE_Rogues_Mod.java
@@ -3,19 +3,18 @@ package com.lne_rogues;
 import com.lne_rogues.config.Default;
 import com.lne_rogues.effect.Effects;
 import com.lne_rogues.item.WeaponRegister;
-import net.fabricmc.api.ModInitializer;
 import com.lne_rogues.config.TweaksConfig;
 import net.fabricmc.loader.api.FabricLoader;
 import net.spell_engine.api.config.ConfigFile;
-import net.tinyconfig.ConfigManager;
+import net.tiny_config.ConfigManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LNE_Rogues_Mod implements ModInitializer {
+public class LNE_Rogues_Mod{
 	public static final String MOD_ID = "lne_rogues";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static ConfigManager<ConfigFile.Equipment> itemConfig = new ConfigManager<>
-			("items", Default.itemConfig)
+			("items_v1", Default.itemConfig)
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)
@@ -27,14 +26,18 @@ public class LNE_Rogues_Mod implements ModInitializer {
 			.sanitize(true)
 			.build();
 
-	@Override
-	public void onInitialize() {
+	public static void init() {
 		tweaksConfig.refresh();
-		Effects.register();
+
+	}
+	public static void registerItems(){
 		if(FabricLoader.getInstance().isModLoaded("loot_n_explore")) {
 			itemConfig.refresh();
 			WeaponRegister.register(itemConfig.value.weapons);
 			itemConfig.save();
 		}
+	}
+	public static void registerEffects(){
+		Effects.register();
 	}
 }