+++ common/src/main/java/com/lne_rogues/client/LNE_Rogues_Client.java
@@ -1,8 +1,5 @@
 package com.lne_rogues.client;
 
-import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.minecraft.util.Identifier;
 import net.spell_engine.api.render.CustomModels;
 
@@ -10,11 +7,9 @@ import java.util.List;
 
 import static com.lne_rogues.LNE_Rogues_Mod.MOD_ID;
 
-@Environment(EnvType.CLIENT)
-public class LNE_Rogues_Client implements ClientModInitializer {
+public class LNE_Rogues_Client {
 
-    @Override
-    public void onInitializeClient() {
+    public static void init() {
         CustomModels.registerModelIds(List.of(
                 Identifier.of(MOD_ID, "projectile/dancing_dagger")
         ));