+++ neoforge/src/main/java/com/lne_rogues/neoforge/client/NeoForgeClient.java
@@ -1,0 +1,20 @@
+package com.lne_rogues.neoforge.client;
+
+import com.lne_rogues.LNE_Rogues_Mod;
+import com.lne_rogues.client.LNE_Rogues_Client;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModLoadingContext;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.spell_engine.client.gui.ConfigMenuScreen;
+
+@EventBusSubscriber(modid = LNE_Rogues_Mod.MOD_ID, value = Dist.CLIENT)
+public class NeoForgeClient {
+    @SubscribeEvent
+    public static void onClientSetup(FMLClientSetupEvent event) {
+        LNE_Rogues_Client.init();
+        ModLoadingContext.get().registerExtensionPoint(IConfigScreenFactory.class, () -> (modContainer, parent) -> new ConfigMenuScreen(parent));
+    }
+}